Added link test and fixed some issues with OneValue and ZeroValue.

+++ shared/src/main/java/com/neuralnetwork/shared/links/ILink.java
@@ -94,6 +94,6 @@ public interface ILink {
      * @return 
      *      the age of this link
      */
-    double getAge();
+    int getAge();
 
 }
